+++ index.js
@@ -1,9 +1,9 @@
 
 import React from 'react';
 import ReactDOM from 'react-dom';
-import Hello from './src/hello';
 import ClusterToggleView from './src/ClusterToggleView';
-import Pong from 'react-pong';
+
+//import ClusterToggleView from './src/ClusterToggleView';
 
 ReactDOM.render(
   <ClusterToggleView data="Juan"></ClusterToggleView>,
