+++ src/ClusterThumbView/ClusterThumbView.jsx
