+++ src/ClusterThumbView/index.js
@@ -1,33 +1,2 @@
-import React from 'react';
-
-
-//Default State
-
-
-class ClusterThumbView extends React.Component{
-  constructor(props) {
-    super(props);
-
-    //Create event binding here
-  }
-
-  render() {
-    return (
-      <div>
-        Hello {this.props.data}
-      </div>
-            );
-  }
-}
-
-
-
-
-//Events management
-
-//Display Data
-
-//Template?
-
-
+import ClusterThumbView from './ClusterThumbView.jsx';
 export default ClusterThumbView
