--- src/hello.js
@@ -1,7 +1,0 @@
-import React from 'react';
-
-export default React.createClass({
-  render: function() {
-    return <div>Hello {this.props.name}</div>;
-  }
-});
